Hackeo para sincronizar los dpoints. Algo de memoria ahorrada por ahí. Arreglados algunos fallos aquí y allí. Comenzado el guardado en logs.

+++ src/com/electrolites/util/SamplePoint.java
@@ -2,9 +2,9 @@ package com.electrolites.util;
 
 public class SamplePoint {
 	public int id;
-	public short sample;
+	public long sample;
 	
-	public SamplePoint(int id, short value) {
+	public SamplePoint(int id, long value) {
 		this.id = id;
 		this.sample = value;
 	}
